Conflicts:
	sli/ingestion/ingestion-base/src/main/java/org/slc/sli/ingestion/landingzone/ZipFileUtil.java

+++ sli/api/src/main/java/org/slc/sli/api/resources/security/RealmResource.java
@@ -14,7 +14,6 @@
  * limitations under the License.
  */
 
-
 package org.slc.sli.api.resources.security;
 
 import java.util.ArrayList;
@@ -63,10 +62,10 @@ import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Component;
 
 /**
- * Realm role mapping API. Allows full CRUD on realm objects. Primarily intended to allow
- * mappings between SLI roles and client roles as realms should not be created or deleted
- * frequently.
- *
+ * Realm role mapping API. Allows full CRUD on realm objects. Primarily intended
+ * to allow mappings between SLI roles and client roles as realms should not be
+ * created or deleted frequently.
+ * 
  * @author jnanney
  */
 @Component
@@ -80,6 +79,8 @@ public class RealmResource {
     public static final String RESPONSE = "response";
     public static final String NAME = "name";
     public static final String UNIQUE_IDENTIFIER = "uniqueIdentifier";
+    public static final String IDP_ID = "idp.id";
+    
     @Autowired
     private EntityDefinitionStore store;
 
@@ -98,17 +99,10 @@ public class RealmResource {
     @Autowired
     private RoleInitializer roleInitializer;
 
-
-
     @PostConstruct
     public void init() {
         EntityDefinition def = store.lookupByResourceName(REALM);
         setService(def.getService());
-    }
-
-    // Injector
-    public void setStore(EntityDefinitionStore store) {
-        this.store = store;
     }
 
     // Injector
@@ -119,7 +113,7 @@ public class RealmResource {
     @PUT
     @Path("{realmId}")
     @Consumes("application/json")
-    @RightsAllowed({Right.CRUD_REALM })
+    @RightsAllowed({ Right.CRUD_REALM })
     public Response updateRealm(@PathParam("realmId") String realmId, EntityBody updatedRealm,
             @Context final UriInfo uriInfo) {
 
@@ -135,7 +129,9 @@ public class RealmResource {
             body.put(RESPONSE, "You are not authorized to update this realm.");
             return Response.status(Status.FORBIDDEN).entity(body).build();
         }
-        Response validateUniqueness = validateUniqueId(realmId, (String) updatedRealm.get(UNIQUE_IDENTIFIER), (String) updatedRealm.get(NAME));
+        Response validateUniqueness = validateUniqueId(realmId, (String) updatedRealm.get(UNIQUE_IDENTIFIER),
+                (String) updatedRealm.get(NAME),
+                getIdpId(updatedRealm));
         if (validateUniqueness != null) {
             return validateUniqueness;
         }
@@ -145,8 +141,8 @@ public class RealmResource {
         updatedRealm.put(ED_ORG, SecurityUtil.getEdOrg());
 
         if (service.update(realmId, updatedRealm)) {
-            audit(securityEventBuilder.createSecurityEvent(RealmResource.class.getName(), uriInfo.getRequestUri(), "Realm ["
-                    + updatedRealm.get(NAME) + "] updated!"));
+            audit(securityEventBuilder.createSecurityEvent(RealmResource.class.getName(), uriInfo.getRequestUri(),
+                    "Realm [" + updatedRealm.get(NAME) + "] updated!"));
             return Response.status(Status.NO_CONTENT).build();
         }
         return Response.status(Status.BAD_REQUEST).build();
@@ -154,18 +150,23 @@ public class RealmResource {
 
     @DELETE
     @Path("{realmId}")
-    @RightsAllowed({Right.CRUD_REALM })
+    @RightsAllowed({ Right.CRUD_REALM })
     public Response deleteRealm(@PathParam("realmId") String realmId, @Context final UriInfo uriInfo) {
         EntityBody deletedRealm = service.get(realmId);
+        if (!canEditCurrentRealm(deletedRealm)) {
+            EntityBody body = new EntityBody();
+            body.put(RESPONSE, "You are not authorized to delete a realm for another ed org");
+            return Response.status(Status.FORBIDDEN).entity(body).build();
+        }
         service.delete(realmId);
         roleInitializer.dropRoles(realmId);
-        audit(securityEventBuilder.createSecurityEvent(RealmResource.class.getName(), uriInfo.getRequestUri(), "Realm ["
-                + deletedRealm.get(NAME) + "] deleted!"));
+        audit(securityEventBuilder.createSecurityEvent(RealmResource.class.getName(), uriInfo.getRequestUri(),
+                "Realm [" + deletedRealm.get(NAME) + "] deleted!"));
         return Response.status(Status.NO_CONTENT).build();
     }
 
     @POST
-    @RightsAllowed({Right.CRUD_REALM })
+    @RightsAllowed({ Right.CRUD_REALM })
     public Response createRealm(EntityBody newRealm, @Context final UriInfo uriInfo) {
 
         if (!canEditCurrentRealm(newRealm)) {
@@ -173,7 +174,8 @@ public class RealmResource {
             body.put(RESPONSE, "You are not authorized to create a realm for another ed org");
             return Response.status(Status.FORBIDDEN).entity(body).build();
         }
-        Response validateUniqueness = validateUniqueId(null, (String) newRealm.get(UNIQUE_IDENTIFIER), (String) newRealm.get(NAME));
+        Response validateUniqueness = validateUniqueId(null, (String) newRealm.get(UNIQUE_IDENTIFIER),
+                (String) newRealm.get(NAME), getIdpId(newRealm));
         if (validateUniqueness != null) {
             debug("On realm create, uniqueId is not unique");
             return validateUniqueness;
@@ -185,11 +187,11 @@ public class RealmResource {
 
         String id = service.create(newRealm);
 
-        //Also create custom roles
+        // Also create custom roles
         roleInitializer.dropAndBuildRoles(id);
 
-        audit(securityEventBuilder.createSecurityEvent(RealmResource.class.getName(), uriInfo.getRequestUri(), "Realm ["
-                + newRealm.get(NAME) + "] created!"));
+        audit(securityEventBuilder.createSecurityEvent(RealmResource.class.getName(), uriInfo.getRequestUri(),
+                "Realm [" + newRealm.get(NAME) + "] created!"));
         String uri = uriToString(uriInfo) + "/" + id;
 
         return Response.status(Status.CREATED).header("Location", uri).build();
@@ -197,19 +199,20 @@ public class RealmResource {
 
     @GET
     @Path("{realmId}")
-    @RightsAllowed({Right.ADMIN_ACCESS })
+    @RightsAllowed({ Right.ADMIN_ACCESS })
     public Response readRealm(@PathParam("realmId") String realmId) {
         EntityBody result = service.get(realmId);
         return Response.ok(result).build();
     }
 
     @GET
-    @RightsAllowed({Right.ADMIN_ACCESS })
+    @RightsAllowed({ Right.ADMIN_ACCESS })
     public Response getRealms(@QueryParam(REALM) @DefaultValue("") String realm, @Context UriInfo info) {
         SLIPrincipal principal = (SLIPrincipal) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
 
         NeutralQuery neutralQuery = new NeutralQuery();
-        neutralQuery.addCriteria(new NeutralCriteria("tenantId", NeutralCriteria.OPERATOR_EQUAL, principal.getTenantId()));
+        neutralQuery.addCriteria(new NeutralCriteria("tenantId", NeutralCriteria.OPERATOR_EQUAL, principal
+                .getTenantId()));
         neutralQuery.addCriteria(new NeutralCriteria(ED_ORG, NeutralCriteria.OPERATOR_EQUAL, principal.getEdOrg()));
         neutralQuery.setOffset(0);
         neutralQuery.setLimit(100);
@@ -235,24 +238,20 @@ public class RealmResource {
         return Response.ok(result).build();
     }
 
-
-    private Response validateUniqueId(String realmId, String uniqueId, String displayName) {
-        if (uniqueId == null || uniqueId.length() == 0) {
-            return null;
-        }
+    private Response validateUniqueId(String realmId, String uniqueId, String displayName, String idpId) {
         // Check for uniqueness of Unique ID
         final NeutralQuery query = new NeutralQuery();
         query.addCriteria(new NeutralCriteria(UNIQUE_IDENTIFIER, "=", uniqueId));
         if (realmId != null) {
             query.addCriteria(new NeutralCriteria("_id", "!=", idConverter.toDatabaseId(realmId)));
         }
-        Entity body =
-                SecurityUtil.runWithAllTenants(new SecurityTask<Entity>() {
+        Entity body = SecurityUtil.runWithAllTenants(new SecurityTask<Entity>() {
 
-                    @Override
-                    public Entity execute() {
-                        return repo.findOne(REALM, query);
-                    } });
+            @Override
+            public Entity execute() {
+                return repo.findOne(REALM, query);
+            }
+        });
 
         if (body != null) {
             debug("uniqueId: {}", body.getBody().get(UNIQUE_IDENTIFIER));
@@ -281,6 +280,27 @@ public class RealmResource {
             res.put(RESPONSE, "Cannot have duplicate display names");
             return Response.status(Status.BAD_REQUEST).entity(res).build();
         }
+        
+        // Check for uniqueness of idp id
+        final NeutralQuery idpIdQuery = new NeutralQuery();
+        idpIdQuery.addCriteria(new NeutralCriteria(IDP_ID, "=", idpId));
+        if (realmId != null) {
+            idpIdQuery.addCriteria(new NeutralCriteria("_id", "!=", idConverter.toDatabaseId(realmId)));
+        }
+        entity = SecurityUtil.runWithAllTenants(new SecurityTask<Entity>() {
+
+            @Override
+            public Entity execute() {
+                return repo.findOne(REALM, idpIdQuery);
+            }
+        });
+
+        if (entity != null) {
+            debug("idp info: {}", getIdpId(entity.getBody()));
+            Map<String, String> res = new HashMap<String, String>();
+            res.put(RESPONSE, "Cannot have duplicate idp ids");
+            return Response.status(Status.BAD_REQUEST).entity(res).build();
+        }
 
         return null;
     }
@@ -292,6 +312,10 @@ public class RealmResource {
     private boolean canEditCurrentRealm(EntityBody realm) {
         String edOrg = SecurityUtil.getEdOrg();
         return !(edOrg == null || !edOrg.equals(realm.get(ED_ORG)));
-
+    }
+    
+    private String getIdpId(Map body) {
+        Map idp = (Map) body.get("idp");
+        return (String) idp.get("id");
     }
 }
